+++ packages/common/src/hooks/index.ts
@@ -24,3 +24,4 @@ export * from './useUSDCBalance'
 export * from './purchaseContent'
 export * from './useAccessAndRemixSettings'
 export * from './useInterval'
+export * from './useCreateUserbankIfNeeded'
