+++ packages/common/src/utils/wallet.ts
@@ -215,10 +215,14 @@ export const formatUSDCWeiToFloorCentsNumber = (amount: BNUSDC) => {
 }
 
 /** General Wallet Utils */
-export const shortenSPLAddress = (addr: string, numChars: number = 4) => {
-  return `${addr.substring(0, numChars)}...${addr.substr(addr.length - numChars - 1)}`
+export const shortenSPLAddress = (addr: string, numChars = 4) => {
+  return `${addr.substring(0, numChars)}...${addr.substr(
+    addr.length - numChars - 1
+  )}`
 }
 
-export const shortenEthAddress = (addr: string, numChars: number = 4) => {
-  return `0x${addr.substring(2, numChars)}...${addr.substr(addr.length - numChars - 1)}`
+export const shortenEthAddress = (addr: string, numChars = 4) => {
+  return `0x${addr.substring(2, numChars)}...${addr.substr(
+    addr.length - numChars - 1
+  )}`
 }
